move map obj to utils file
https://coveord.atlassian.net/browse/KIT-2600
